+++ crates/metadata/src/lib.rs
@@ -2,7 +2,8 @@ mod git;
 mod metadata;
 mod metadata_error;
 pub use metadata::{
-    Build, ClockType, FilelistType, Format, Metadata, PathPair, Project, ResetType, Target,
+    Build, BuiltinType, ClockType, FilelistType, Format, Metadata, PathPair, Project, ResetType,
+    Target,
 };
 pub use metadata_error::MetadataError;
 pub use semver;